Fix memory-tracking
As identified in https://github.com/sine-fdn/polytune/pull/73#issuecomment-2944644105
there is a problem with the previous iteration of memory tracking
when having nested tokio::spawns.
This commit fixes the issue by using separate tokio runtimes for each
party, where on thread start, the PARTY_IDX is set correctly for each
worker thread.
Incidentally, this also increases the performance of the runtime
benchmarks.