Concurrent sending/receiving in src/fpre.rs
As part of #61, this changes the trusted dealer implementation in
src/fpre.rs to use concurrent sending and receiving.

The first part of this commit was written by me and finished by claude
code with the following prompt:
In @src/fpre.rs there are some cases left where send and recv methods
that are await are used in loop. Instead we want to do it concurrently.
I've already refactored some of the uses to try_join_all. Finish the
remaining cases where async methods are used in a loop.