Concurrent sending/receiving in src/faand.rs
Part of the work on #61. This introduces concurrent sending and
receving in the src/faand.rs implementation of preprocessing.

Note: This commit was created with the help of claude code in an
interactive session. I instructed claude to change the loops containing
awaited futures similar to src/protocol.rs and src/fpre.rs. This
required some iterations.