Add crates/polytune-server-core
This commit adds the polytune-server-core crate. It implements the
execution of a Policy as a state machine which is executed as a future
and communicated with using an mpsc::channel. The state machine only
receives PoliyCmds and communicates witht the other state machines
via a PolicyClient trait. This trait can be implemented in different
ways, e.g., using an HTTP API as in the previous polytune-server.