http-server: compile with no RUSTFLAGS
If the CI runner that compiles the http-server image has a different
instruction set than the host that later runs the server, this
can result in a SIGILL. This happened as by default the RUSTFLAGS
configured in .cargo/config.toml were used.

This commit fixes that by explicitly setting the RUSTFLAGS to "".
The server will still be able to use aes and avx2 instructions for
encryption and bitmatrix transpose, as these use runtime feature
detection.