CI: Add http-server image tagging
This adds image tagging for the http-server docker image on pushed
git tags. If a git tag with a `http-server-v` prefix is pushed,
the version is used as the tag for the docker image. The docker
`latest` tag points to the latest tagged release. The `dev` tag
corresponds to the latest push to the main branch.