CI: Add cooldown to dependabot updates
Adding a cooldown to dependency updates can reduce the risk that we
fall victim to a supply chain attacks due to a compromised dependency.
In many cases, there is a somewhat short window of opportunity after
a malicious version has been published and this is uncovered. By adding,
a cooldown, we hopefully fall outside of this window of opportunity of
the attacker.