http-server: change URL handling
This changes the URL handling of the policy.participants to be
independent of a trailing slash or not. Previously, if a participant
had a URL like `http://localhost:8000/some-path` (no trailing slash),
the server would replace the `some-path` part with the RPC endponts
like `validate` instead of appending a new segment after `some-path`.
This was due to documented behavior of the `Url::join` method, but this
caused confusion for us (and others, see https://github.com/servo/rust-url/issues/333).

With this change, there is no difference between a participant URL having
a trailing slash or not.