polytune: critical fix, remove fixed seed for OTs
PR #202 introduced a critical issue in the implementation of the KOS OTs.
The KOS OT sender was always initialized with a PRG from a fixed seed
(value 42). If I'm not mistaken, this breaks the complete security of
the protocol, as the OTs are not secure anymore.

Why did this happen?: I believe I made this change when debugging the
register based circuit evaluation of #202. There were some randomly
occuring errors, so I tried to remove sources of non-determinism to
better debug the issue. It seems, this one change slipped through.